+++ lib/commons/color/get-text-shadow-colors.js
@@ -1,13 +1,7 @@
 import Color from './color';
 import assert from '../../core/utils/assert';
-
-/** Magic numbers **/
-// Alpha value to use when text shadows are offset between .5px and 1.5px
-const SHADOW_STROKE_ALPHA = 0.54;
-// Shadows offset by less than this are not visible enough no matter how much you stack them
-const VISIBLE_SHADOW_MIN_PX = 0.5;
-// Shadows offset by more than this have full opacity
-const OPAQUE_STROKE_OFFSET_MIN_PX = 1.5;
+import getStrokeColorsFromShadows from './get-stroke-colors-from-shadows';
+import parseTextShadows from './parse-text-shadows';
 
 /**
  * Get text-shadow colors that can impact the color contrast of the text, including from:
@@ -48,7 +42,7 @@ export default function getTextShadowColors(node, { minRatio, maxRatio } = {}) {
     }
     if (thinShadows.length > 0) {
       // Inset any stroke colors before this shadow
-      const strokeColors = getStrokeColor(thinShadows);
+      const strokeColors = getStrokeColorsFromShadows(thinShadows);
       shadowColors.push(...strokeColors);
       thinShadows.splice(0, thinShadows.length); // empty
     }
@@ -65,115 +59,11 @@ export default function getTextShadowColors(node, { minRatio, maxRatio } = {}) {
 
   if (thinShadows.length > 0) {
     // Append any remaining stroke colors
-    const strokeColors = getStrokeColor(thinShadows);
+    const strokeColors = getStrokeColorsFromShadows(thinShadows);
     shadowColors.push(...strokeColors);
   }
 
   return shadowColors;
-}
-
-function getStrokeColor(thinShadows) {
-  const edges = ['top', 'right', 'bottom', 'left'];
-  const colorMap = {};
-  for (const { colorStr, pixels } of thinShadows) {
-    colorMap[colorStr] ??= { top: [], right: [], bottom: [], left: [] };
-    const borders = colorMap[colorStr];
-    const [offsetX, offsetY] = pixels;
-
-    if (offsetX > VISIBLE_SHADOW_MIN_PX) {
-      borders.right.push(pixels);
-    } else if (offsetX < -VISIBLE_SHADOW_MIN_PX) {
-      borders.left.push(pixels);
-    }
-    if (offsetY > VISIBLE_SHADOW_MIN_PX) {
-      borders.bottom.push(pixels);
-    } else if (offsetY < -VISIBLE_SHADOW_MIN_PX) {
-      borders.top.push(pixels);
-    }
-  }
-
-  const strokeShadows = Object.entries(colorMap).filter(([, sides]) => {
-    const sidesCovered = edges.filter(side => sides[side].length !== 0).length;
-    return sidesCovered >= 3;
-  });
-
-  const strokeColors = strokeShadows.map(([colorStr, sides]) => {
-    const strokeColor = new Color();
-    strokeColor.parseString(colorStr);
-
-    // Average the number of shadows around the text
-    let density = 0;
-    let isSolid = true;
-    edges.forEach(edge => {
-      density += sides[edge].length / 4;
-      isSolid &&= sides[edge].every(
-        ([x, y]) =>
-          Math.abs(x) > OPAQUE_STROKE_OFFSET_MIN_PX ||
-          Math.abs(y) > OPAQUE_STROKE_OFFSET_MIN_PX
-      );
-    });
-
-    if (!isSolid) {
-      // As more shadows surround the text, the opacity increases
-      strokeColor.alpha = 1 - Math.pow(SHADOW_STROKE_ALPHA, density);
-    }
-    return strokeColor;
-  });
-
-  return strokeColors;
-}
-
-/**
- * Parse text-shadow property value. Required for IE, which can return the color
- * either at the start or the end, and either in rgb(a) or as a named color
- */
-function parseTextShadows(textShadow) {
-  let current = { pixels: [] };
-  let str = textShadow.trim();
-  const shadows = [current];
-  if (!str) {
-    return [];
-  }
-
-  while (str) {
-    const colorMatch =
-      str.match(/^rgba?\([0-9,.\s]+\)/i) ||
-      str.match(/^[a-z]+/i) ||
-      str.match(/^#[0-9a-f]+/i);
-    const pixelMatch = str.match(/^([0-9.-]+)px/i) || str.match(/^(0)/);
-
-    if (colorMatch) {
-      assert(
-        !current.colorStr,
-        `Multiple colors identified in text-shadow: ${textShadow}`
-      );
-      str = str.replace(colorMatch[0], '').trim();
-      current.colorStr = colorMatch[0];
-    } else if (pixelMatch) {
-      assert(
-        current.pixels.length < 3,
-        `Too many pixel units in text-shadow: ${textShadow}`
-      );
-      str = str.replace(pixelMatch[0], '').trim();
-      const pixelUnit = parseFloat(
-        (pixelMatch[1][0] === '.' ? '0' : '') + pixelMatch[1]
-      );
-      current.pixels.push(pixelUnit);
-    } else if (str[0] === ',') {
-      // multiple text-shadows in a single string (e.g. `text-shadow: 1px 1px 1px #000, 3px 3px 5px blue;`
-      assert(
-        current.pixels.length >= 2,
-        `Missing pixel value in text-shadow: ${textShadow}`
-      );
-      current = { pixels: [] };
-      shadows.push(current);
-      str = str.substr(1).trim();
-    } else {
-      throw new Error(`Unable to process text-shadows: ${textShadow}`);
-    }
-  }
-
-  return shadows;
 }
 
 function textShadowColor({ colorStr, offsetX, offsetY, blurRadius, fontSize }) {
