+++ lib/commons/color/index.js
@@ -16,7 +16,9 @@ export { default as getContrast } from './get-contrast';
 export { default as getForegroundColor } from './get-foreground-color';
 export { default as getOwnBackgroundColor } from './get-own-background-color';
 export { default as getRectStack } from './get-rect-stack';
+export { default as getStrokeColorsFromShadows } from './get-stroke-colors-from-shadows';
+export { default as getTextShadowColors } from './get-text-shadow-colors';
 export { default as hasValidContrastRatio } from './has-valid-contrast-ratio';
 export { default as incompleteData } from './incomplete-data';
-export { default as getTextShadowColors } from './get-text-shadow-colors';
+export { default as parseTextShadows } from './parse-text-shadows';
 export { getStackingContext, stackingContextToColor } from './stacking-context';
