+++ test/commons/color/get-text-shadow-colors.js
@@ -231,103 +231,6 @@ describe('axe.commons.color.getTextShadowColors', function () {
       assert.lengthOf(shadowColors, 4);
     });
 
-    it('skips raised-letter effects (shadows on 1 or 2 sides)', () => {
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          0 -1px #000,
-          -1px -1px #000,
-          -1px 0 #000,
-          1px 1px #fff;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 0);
-    });
-
-    it('skips shadows offset with 0.5px or less', () => {
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          0.5px 0.5px #000,
-          -0.5px -0.5px #000;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 0);
-    });
-
-    it('applies an alpha value to shadows of 1.5px or less', () => {
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          0 -1.5px #000,
-          1.5px 0 #000,
-          0 1.5px #000,
-          -1.5px 0 #000;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 1);
-      assert.equal(shadowColors[0].red, 0);
-      assert.equal(shadowColors[0].green, 0);
-      assert.equal(shadowColors[0].blue, 0);
-      assert.closeTo(shadowColors[0].alpha, 0.46, 0.01);
-    });
-
-    it('applies an alpha value if not all shadows are offset by more than 1.5px', () => {
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          0 -2px #000,
-          2px 0 #000,
-          0 1.5px #000,
-          -1.5px 0 #000;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 1);
-      assert.equal(shadowColors[0].red, 0);
-      assert.equal(shadowColors[0].green, 0);
-      assert.equal(shadowColors[0].blue, 0);
-      assert.closeTo(shadowColors[0].alpha, 0.46, 0.01);
-    });
-
-    it('multiplies alpha value for each shadow', () => {
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          0 -1px #000,
-          0 -1px #000,
-          1px 0 #000,
-          1px 0 #000,
-          0 1px #000,
-          0 1px #000,
-          -1px 0 #000,
-          -1px 0 #000;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 1);
-      assert.equal(shadowColors[0].red, 0);
-      assert.equal(shadowColors[0].green, 0);
-      assert.equal(shadowColors[0].blue, 0);
-      assert.closeTo(shadowColors[0].alpha, 0.7, 0.01);
-    });
-
-    it('double-counts shadows on corners', () => {
-      // Corner-shadows overlap on the sides of letters, increasing alpha
-      fixture.innerHTML = `
-        <span style="text-shadow:
-          -1px -1px #000,
-          1px -1px #000,
-          1px 1px #000,
-          -1px 1px #000;
-        ">Hello wold</span>
-      `;
-      const shadowColors = getTextShadowColors(fixture.firstElementChild, opt);
-      assert.lengthOf(shadowColors, 1);
-      assert.equal(shadowColors[0].red, 0);
-      assert.equal(shadowColors[0].green, 0);
-      assert.equal(shadowColors[0].blue, 0);
-      assert.closeTo(shadowColors[0].alpha, 0.7, 0.01);
-    });
-
     it('places the combined shadow in the correct order with shadows', () => {
       fixture.innerHTML = `
         <span style="text-shadow:
